change src/common to integer8, test passed for src/common

+++ src/common/incoming_fields.f90
@@ -5,7 +5,7 @@ subroutine point_source_scalar_helmholtz(P0,ns,points,normals,zk,pot,dpot_dnorma
 implicit none
 
 	!List of calling arguments
-	integer ( kind = 4 ), intent(in)  :: ns
+	integer ( kind = 8 ), intent(in)  :: ns
 	real ( kind = 8 ), intent(in) :: P0(3)
 	real ( kind = 8 ), intent(in) :: points(3,ns)
 	real ( kind = 8 ), intent(in) :: normals(3,ns)
@@ -14,7 +14,7 @@ implicit none
 	complex ( kind = 8 ), intent(out) :: dpot_dnormal(ns)
 
     !List of local variables
-	integer ( kind = 4 ) count1
+	integer ( kind = 8 ) count1
 	real ( kind = 8 ) r
 	complex (kind = 8 ) ima
 	complex ( kind = 8 ) gradpot_aux(3),aux
@@ -39,7 +39,7 @@ subroutine point_source_scalar_helmholtz2(P0,qv,ns,points,zk,pot,gradpot,init)
 implicit none
 
 	!List of calling arguments
-	integer ( kind = 4 ), intent(in)  :: ns,init
+	integer ( kind = 8 ), intent(in)  :: ns,init
 	real ( kind = 8 ), intent(in) :: P0(3)
 	real ( kind = 8 ), intent(in) :: points(3,ns)
 	complex ( kind = 8 ), intent(in) :: zk,qv
@@ -47,7 +47,7 @@ implicit none
 	complex ( kind = 8 ), intent(out) :: gradpot(3,ns)
 
     !List of local variables
-	integer ( kind = 4 ) count1
+	integer ( kind = 8 ) count1
 	real ( kind = 8 ) r
 	complex (kind = 8 ) ima
 	complex ( kind = 8 ) aux
@@ -77,7 +77,7 @@ implicit none
 ! a small electric dipole; that is: E=vf*exp(ima*zk*r)/r 
 
 	!List of calling arguments
-	integer ( kind = 4 ), intent(in) :: ns
+	integer ( kind = 8 ), intent(in) :: ns
 	real ( kind = 8 ), intent(in) :: P0(3)
 	complex ( kind = 8 ), intent(in) :: vf(3)
 	real ( kind = 8 ), intent(in) :: points(3,ns)
@@ -87,7 +87,7 @@ implicit none
 	complex ( kind = 8 ), intent(out) :: divE(ns)
 	
     !List of local variables
-	integer ( kind = 4 ) count1
+	integer ( kind = 8 ) count1
 	real ( kind = 8 ) dx,dy,dz,r
 	complex (kind = 8 ) ima
 	complex (kind = 8 ) R1
@@ -127,7 +127,7 @@ subroutine fieldsED(zk, P0, points, n, E, H, vf, init)
   ! where  A = vf*exp(ima*zk*r)/r
 	
   ! List of calling arguments
-  integer, intent(in) :: n,init
+  integer(8), intent(in) :: n,init
   real ( kind = 8 ), intent(in) :: P0(3),points(12,n)
   complex ( kind = 8 ), intent(in) :: vf(3),zk
   complex ( kind = 8 ), intent(out) :: E(3,n),H(3,n)
@@ -135,7 +135,7 @@ subroutine fieldsED(zk, P0, points, n, E, H, vf, init)
   ! List of local variables
   real ( kind = 8 ) dx,dy,dz,r
   complex ( kind = 8 ) R1,R2,au1,au2,ima
-  integer i
+  integer(8) i
 	
   ima = (0.0d0,1.0d0)
   do i=1,n
@@ -184,7 +184,7 @@ implicit none
 ! H=1/(ima*zk)*curlcurlF
 
 	!List of calling arguments
-	integer, intent(in) :: n,initial
+	integer(8), intent(in) :: n,initial
 	real ( kind = 8 ), intent(in) :: P0(3),points(12,n)
 	complex ( kind = 8 ), intent(in) :: vf(3),zk
 	complex ( kind = 8 ), intent(out) :: E(3,n),H(3,n)
@@ -192,7 +192,7 @@ implicit none
 	!List of local variables
 	real ( kind = 8 ) dx,dy,dz,r
 	complex ( kind = 8 ) R1,R2,au1,au2,ima
-	integer i
+	integer(8) i
 
 	ima = (0.0d0,1.0d0)
 
@@ -234,12 +234,12 @@ subroutine fieldsPW(zk,xyz,n,nsource,E,H)
 ! a plane wave; that is: E=exp(ima*zk*z) ux
 ! H=exp(ima*zk*z) uy
 
-	integer, intent(in) :: n
+	integer(8), intent(in) :: n
 	real ( kind = 8 ), intent(in) :: xyz(3,n)
 	complex ( kind = 8 ), intent(in) :: zk
 	complex *16 E(3,n),H(3,n)
 	complex *16 ima
-	integer i
+	integer(8) i
 	data ima/(0.0d0,1.0d0)/
 
       do i=1,n
@@ -262,13 +262,13 @@ subroutine ScalarPW(zk,xyz,n,pot)
 ! a plane wave; that is: E=exp(ima*zk*z) ux
 ! H=exp(ima*zk*z) uy
 
-	integer, intent(in) :: n
+	integer(8), intent(in) :: n
 	real ( kind = 8 ), intent(in) :: xyz(3,n)
 	complex ( kind = 8 ), intent(in) :: zk
 	complex *16 pot(n)
 	complex *16 ima
 	real ( kind = 8 ) phi,pheta,kx,ky,kz
-	integer i
+	integer(8) i
 	data ima/(0.0d0,1.0d0)/
 	real ( kind = 8 ) pi
 	
@@ -349,7 +349,7 @@ implicit none
 ! E=1/(-ima*omega*ep)*curlcurlA
 	
 	!List of calling arguments
-	integer, intent(in) :: n,init
+	integer(8), intent(in) :: n,init
 	real ( kind = 8 ), intent(in) :: P0(3),points(12,n)
 	complex ( kind = 8 ), intent(in) :: vf(3),omega,ep,mu
 	complex ( kind = 8 ), intent(out) :: E(3,n),H(3,n)
@@ -357,7 +357,7 @@ implicit none
 	!List of local variables
 	real ( kind = 8 ) dx,dy,dz,r
 	complex ( kind = 8 ) R1,R2,au1,au2,ima,zk
-	integer i
+	integer(8) i
 	
 	ima = (0.0d0,1.0d0)
 	zk=omega*sqrt(ep*mu)
@@ -400,7 +400,7 @@ implicit none
 ! H=1/(ima*omega*mu)*curlcurlF
 
 	!List of calling arguments
-	integer, intent(in) :: n,initial
+	integer(8), intent(in) :: n,initial
 	real ( kind = 8 ), intent(in) :: P0(3),points(12,n)
 	complex ( kind = 8 ), intent(in) :: vf(3),omega,ep,mu
 	complex ( kind = 8 ), intent(out) :: E(3,n),H(3,n)
@@ -408,7 +408,7 @@ implicit none
 	!List of local variables
 	real ( kind = 8 ) dx,dy,dz,r,zk
 	complex ( kind = 8 ) R1,R2,au1,au2,ima
-	integer i
+	integer(8) i
 	
 	zk=omega*sqrt(ep*mu)
 	ima = (0.0d0,1.0d0)
@@ -455,7 +455,7 @@ implicit none
 ! E=1/(-ima*zk)*curlcurlA
 	
 	!List of calling arguments
-	integer, intent(in) :: n,init,ndtarg
+	integer(8), intent(in) :: n,init,ndtarg
 	real ( kind = 8 ), intent(in) :: P0(3),points(ndtarg,n)
 	complex ( kind = 8 ), intent(in) :: vf(3),zk
 	complex ( kind = 8 ), intent(out) :: E(3,n),H(3,n),A(3,n)
@@ -465,7 +465,7 @@ implicit none
 	real ( kind = 8 ) dx,dy,dz,r
 	complex ( kind = 8 ) R1,R2,au1,au2,ima,myexp
 	complex ( kind = 8 ) e_aux1,e_aux2,e_aux3,a_aux1,a_aux2,a_aux3
-	integer i
+	integer(8) i
 	
 	ima = (0.0d0,1.0d0)
 
@@ -545,7 +545,7 @@ implicit none
 ! H=1/(ima*zk)*curlcurlF
 
 	!List of calling arguments
-	integer, intent(in) :: n,initial,ndtarg
+	integer(8), intent(in) :: n,initial,ndtarg
 	real ( kind = 8 ), intent(in) :: P0(3),points(ndtarg,n)
 	complex ( kind = 8 ), intent(in) :: vf(3),zk
 	complex ( kind = 8 ), intent(out) :: E(3,n),H(3,n),A(3,n)
@@ -554,7 +554,7 @@ implicit none
 	!List of local variables
 	real ( kind = 8 ) dx,dy,dz,r
 	complex ( kind = 8 ) R1,R2,au1,au2,ima,e_aux1,e_aux2,e_aux3,myexp
-	integer i
+	integer(8) i
 	
 	ima = (0.0d0,1.0d0)
